Remove functions and variable common to both pokedex from the general pokedex interface to their own file
