+++ src/pokedex.c
@@ -16,6 +16,7 @@
 #include "palette.h"
 #include "pokedex.h"
 #include "pokedex_area_screen.h"
+#include "pokedex_common.h"
 #include "pokedex_cry_screen.h"
 #include "pokedex_plus_hgss.h"
 #include "rtc.h"
